Item and disguise framework
Items are now stored in the inventory, and dropped when another is picked up.

disguise works the same way, but when you press 'F' when holding a disguise, you change into that disguise (or back to default if you are already disguised)

item location raycast is no longer on character script. moved to inventory for ease of access.