Created Warden AI
Warden is inactive until suspicion meter reaches 3. (area1metric = 3)
I had to change the protection level on that float Alex, so its public now, but thats all.

When warden is active, any hack can be thwarted if the player remains in the hack for too long (set time in inspector)