Created a projectile that can be thrown from the player
projectile can be thrown from players position, while stationary and moving, arc is limited to the camera's orintation.
May look into another way to do this.

The Shoot Script is set up to eventually take from the player's inventory, so that player can only throw the object if they have it in their inventory (all comments)