Art assets and Projectile linked to inventory
the inventory works with the projectile as it was set up, but its not really an inventory, this will need to be fixed. (objects aren't added to an inventory, a reference to them is)

Art has been added, mainly the prison cell art. I am not sure if it just my pc but its quite laggy. Fixed some door animations with the new models