Networking bull
player cameras are once again messed up and cannot access majority of their functions. players can still connect and their characters do spawn, however Dustin cannot be viewed, and the minimap works on and off.

If the player or Car already in the scene is removed the security camera system doesn't work either as it needs to be set to the new car but I cannot do that yet