Quest system updates
currently 2 quests exists, find a level 1, then level 2, keycard. Will add more later and update the activating quest system.
press "TAB" to bring up the quest descriptions on the left hand side of the screen.

for the new cube (cart) press "e" to grab and drag, and "e" to let go when looking at it