Added Post Process Need Assistance
Added a post process volume into main scene and a post process layer script to the main camera so it can be effected. However, was not able to create a new Layer for the post process effects and it was recommended that I do not set the Layer to Everything or Default. If someone could look into that one, that'd be great. Also added some cafeteria assets and textures to hallways